requirements.txt
streamlit
scikit-learn
PyPDF2
python-docx
nltk
pandas